core: split large async hot paths out of codex.rs
